refactor: 결제 성공 시 재고 차감 로직 추가 (#15)
