+++ apps/commerce-api/src/main/java/com/loopers/application/payment/handler/CardPaymentHandler.java
@@ -1,10 +1,13 @@
-package com.loopers.application.payment.handler.card;
-
+package com.loopers.application.payment.handler;
 
 import static com.loopers.domain.payment.PaymentProvider.SIMULATOR;
 
 import com.loopers.application.payment.PaymentCommand.PaymentRequestCommand;
 import com.loopers.application.payment.data.CardMethod;
+import com.loopers.application.payment.data.PaymentMethod;
+import com.loopers.domain.catalog.Inventory;
+import com.loopers.domain.catalog.Product;
+import com.loopers.domain.catalog.ProductService;
 import com.loopers.domain.coupon.IssuedCoupon;
 import com.loopers.domain.coupon.IssuedCouponService;
 import com.loopers.domain.order.Order;
@@ -14,37 +17,38 @@ import com.loopers.domain.payment.Payment;
 import com.loopers.domain.payment.PaymentClient;
 import com.loopers.domain.payment.PaymentClientData.PaymentClientRequest;
 import com.loopers.domain.payment.PaymentClientData.PaymentClientResponse;
-import com.loopers.domain.payment.PaymentProvider;
-import com.loopers.domain.payment.SupportPg.PgSimulator;
+import java.util.List;
+import java.util.Map;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+
+@RequiredArgsConstructor
 @Component
-class PaymentGatewaySimulator implements PaymentGateway {
+public class CardPaymentHandler implements PaymentHandler {
 
     @Value("${external.payment.providers.simulator.callback-url}")
     private String callBackUrl;
 
     private final PaymentClient paymentClient;
     private final OrderService orderService;
+    private final ProductService productService;
     private final IssuedCouponService issuedCouponService;
 
-    public PaymentGatewaySimulator(@PgSimulator PaymentClient paymentClient, OrderService orderService,
-        IssuedCouponService issuedCouponService) {
-        this.paymentClient = paymentClient;
-        this.orderService = orderService;
-        this.issuedCouponService = issuedCouponService;
+    @Override
+    public boolean supports(PaymentMethod paymentMethod) {
+        return paymentMethod instanceof CardMethod;
     }
 
     @Override
-    public boolean supports(PaymentProvider selectedPg) {
-        return selectedPg == SIMULATOR;
-    }
-
-    @Override
-    public Payment processPayment(PaymentRequestCommand command) {
+    public Payment process(PaymentRequestCommand command) {
         Order order = orderService.findByOrderNumber(command.orderNumber());
         CardMethod cardData = (CardMethod) command.paymentMethod();
+        List<Product> products = productService.findAllWithLock(order.purchaseProductIds());
+
+        Map<Long, Long> stockMap = order.purchaseProducts();
+        Inventory inventory = Inventory.convert(products, stockMap);
 
         // 결제 요청
         PaymentClientRequest paymentClientRequest = new PaymentClientRequest(order.getOrderNumber().number(), cardData.cardType(),
@@ -63,6 +67,11 @@ class PaymentGatewaySimulator implements PaymentGateway {
             return CardPayment.failCard(order.getOrderNumber().number(), order.paidAmount(), cardData.cardType(), SIMULATOR,
                 paymentClientResponse);
         } else {
+            // 주문 성공
+            order.complete();
+
+            // 재고 차감
+            inventory.decreaseStock();
             return CardPayment.initiate(order.getOrderNumber().number(), order.paidAmount(), cardData.cardType(), SIMULATOR,
                 paymentClientResponse, paymentClientResponse.transactionKey());
         }
