+++ apps/commerce-api/src/main/java/com/loopers/application/payment/handler/PointPaymentHandler.java
@@ -1,9 +1,11 @@
-package com.loopers.application.payment.handler.point;
+package com.loopers.application.payment.handler;
 
 import com.loopers.application.payment.PaymentCommand.PaymentRequestCommand;
 import com.loopers.application.payment.data.PaymentMethod;
 import com.loopers.application.payment.data.PointMethod;
-import com.loopers.application.payment.handler.PaymentHandler;
+import com.loopers.domain.catalog.Inventory;
+import com.loopers.domain.catalog.Product;
+import com.loopers.domain.catalog.ProductService;
 import com.loopers.domain.coupon.IssuedCoupon;
 import com.loopers.domain.coupon.IssuedCouponService;
 import com.loopers.domain.order.Order;
@@ -13,6 +15,8 @@ import com.loopers.domain.payment.PointPayment;
 import com.loopers.domain.shared.ExceptionMatcher;
 import com.loopers.domain.user.User;
 import com.loopers.domain.user.UserService;
+import java.util.List;
+import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -24,6 +28,7 @@ class PointPaymentHandler implements PaymentHandler {
 
     private final OrderService orderService;
     private final UserService userService;
+    private final ProductService productService;
     private final ExceptionMatcher exceptionMatcher;
     private final IssuedCouponService issuedCouponService;
 
@@ -36,6 +41,11 @@ class PointPaymentHandler implements PaymentHandler {
     public Payment process(PaymentRequestCommand command) {
         Order order = orderService.findByOrderNumberWithLock(command.orderNumber());
         User user = userService.findByIdWithLock(order.getBuyerId());
+
+        List<Product> products = productService.findAllWithLock(order.purchaseProductIds());
+
+        Map<Long, Long> stockMap = order.purchaseProducts();
+        Inventory inventory = Inventory.convert(products, stockMap);
 
         // 결제 생성
         PointPayment pointPayment = PointPayment.initiate(order.getOrderNumber().number(), order.paidAmount(), user.getId(),
@@ -50,6 +60,10 @@ class PointPaymentHandler implements PaymentHandler {
 
             // 주문 성공
             order.complete();
+
+            // 재고 차감
+            inventory.decreaseStock();
+
         } catch (Exception e) {
             if (exceptionMatcher.isNotEnoughPoint(e)) {
                 pointPayment.fail("포인트가 부족합니다.");
